Changed module name to 'coinflip'. Cleaned up code. Fixed error in test.
